requirements.txt
selenium~=4.40.0
pytest~=9.0.2
pytest-html
webdriver-manager~=4.0.2
h11~=0.16.0
jupyter~=1.0.0
jupyterlab-pygments~=0.1.2
pip~=21.2.3
attrs~=25.4.0
PySocks~=1.7.1
tornado~=6.1
Jinja2~=3.0.3
filelock~=3.10.0
cffi~=1.15.0
setuptools~=57.4.0
typing_extensions~=4.15.0
trio~=0.33.0
outcome~=1.3.0.post0
async-generator~=1.10
sniffio~=1.3.1
exceptiongroup~=1.3.1
sortedcontainers~=2.4.0
idna~=3.11
ipython~=7.29.0
MarkupSafe~=2.0.1
Pygments~=2.10.0
wcwidth~=0.2.5
pluggy~=1.6.0
iniconfig~=2.3.0
numpy~=1.22.2
certifi~=2026.1.4
urllib3~=2.6.3
charset-normalizer~=3.4.4
trio-websocket~=0.12.2
websocket-client~=1.9.0
wsproto~=1.3.2
requests~=2.32.5
python-dotenv~=1.2.1
packaging~=26.0
allure-python-commons~=2.15.3